gnrc_ipv6_nib/arsm: ensure proper int width in backoff calculation
Otherwise, the result might flow over on 8/16-bit platforms

(cherry picked from commit 1f4d57d608ac62bead410e282d4e14d4085b5e6a)

+++ sys/net/gnrc/network_layer/ipv6/nib/_nib-arsm.c
@@ -465,7 +465,8 @@ static inline uint32_t _exp_backoff_retrans_timer(uint8_t ns_sent,
 
     /* backoff according to  https://tools.ietf.org/html/rfc7048 with
      * BACKOFF_MULTIPLE == 2 */
-    tmp = ((1 << ns_sent) * retrans_timer * tmp) / US_PER_MS;
+    tmp = (uint32_t)(((uint64_t)(((uint32_t) 1) << ns_sent) * retrans_timer *
+                     tmp) / US_PER_MS);
     /* random factors were statically multiplied with 1000 ^ */
     if (tmp > NDP_MAX_RETRANS_TIMER_MS) {
         tmp = NDP_MAX_RETRANS_TIMER_MS;
